+++ src/types/login.types.ts
@@ -11,3 +11,22 @@ export type LoginHookReturnType = [
   LoginHookReturnType,
   LoginHookGenericStateType
 ];
+
+export interface OnLoginRedirectOptionsType {
+  signature?: string;
+  address: string;
+}
+
+export type OnLoginRedirectType = (
+  callbackRoute: string,
+  options?: OnLoginRedirectOptionsType
+) => void;
+
+export interface OnProviderLoginType {
+  callbackRoute?: string;
+  token?: string;
+  /**
+   * If specified, `onLoginRedirect` will overwrite callbackRoute default navigation
+   */
+  onLoginRedirect?: OnLoginRedirectType;
+}