+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[2.1.3](https://github.com/ElrondNetwork/dapp-core/pull/455)] - 2022-10-31
+- [Changed `onLoginRedirect` callback logic to overwrite callbackRoute navigation](https://github.com/ElrondNetwork/dapp-core/pull/454)
+- [Changed `onLoginRedirect` function to allow sending signature and address params](https://github.com/ElrondNetwork/dapp-core/pull/453)
+
 ## [[2.1.2](https://github.com/ElrondNetwork/dapp-core/pull/451)] - 2022-10-27
 
 - [Fixed throwing console exception on searching for transaction](https://github.com/ElrondNetwork/dapp-core/pull/449)